Move Ands and Ors to own files
