+++ community/cypher/cypher-compiler-2.3/src/main/scala/org/neo4j/cypher/internal/compiler/v2_3/commands/predicates/Predicate.scala
@@ -51,114 +51,6 @@ object Predicate {
   def fromSeq(in: Seq[Predicate]) = in.reduceOption(_ andWith _).getOrElse(True())
 }
 
-object And {
-  def apply(a: Predicate, b: Predicate) = (a, b) match {
-    case (True(), other) => other
-    case (other, True()) => other
-    case (_, _)          => new And(a, b)
-  }
-}
-
-object Ands {
-  def apply(a: Predicate, b: Predicate) = (a, b) match {
-    case (True(), other) => other
-    case (other, True()) => other
-    case (_, _)          => new Ands(List(a, b))
-  }
-}
-
-case class Ands(predicates: List[Predicate]) extends CompositeBooleanPredicate {
-  def shouldExitWhen = false
-  override def andWith(other: Predicate): Predicate = Ands(predicates :+ other)
-  def rewrite(f: (Expression) => Expression): Expression = f(Ands(predicates.map(_.rewriteAsPredicate(f))))
-}
-
-case class AndedPropertyComparablePredicates(ident: Identifier, prop: Property, comparables: NonEmptyList[ComparablePredicate]) extends Predicate {
-
-  def symbolTableDependencies: Set[String] = comparables.foldLeft(prop.symbolTableDependencies) {
-    case (acc, predicate) => acc ++ predicate.symbolTableDependencies
-  }
-
-  def containsIsNull: Boolean = comparables.exists(_.containsIsNull)
-
-  def isMatch(m: ExecutionContext)(implicit state: QueryState): Option[Boolean] = {
-    var result: Option[Option[Boolean]] = None
-    val iter = comparables.toIterable.iterator
-    while (iter.nonEmpty) {
-      val p = iter.next()
-      val r = p.isMatch(m)
-
-      if(r.nonEmpty && !r.get)
-        return r
-
-      if(result.isEmpty)
-        result = Some(r)
-      else {
-        val stored = result.get
-        if (stored.nonEmpty && stored.get && r.isEmpty)
-          result = Some(None)
-      }
-    }
-
-    result.get
-  }
-
-  def arguments: Seq[Expression] = comparables.toSeq
-
-  // some rewriters change the type of this, and we can't allow that
-  private def rewriteIdentifierIfNotTypeChanged(f: (Expression) => Expression) =
-    ident.rewrite(f) match {
-      case i: Identifier => i
-      case _ => ident
-    }
-
-  def rewrite(f: (Expression) => Expression): Expression =
-    f(AndedPropertyComparablePredicates(rewriteIdentifierIfNotTypeChanged(f),
-                                        prop.rewrite(f).asInstanceOf[Property],
-                                        comparables.map(_.rewriteAsPredicate(f).asInstanceOf[ComparablePredicate])))
-
-  override def atoms: Seq[Predicate] = comparables.toSeq
-}
-
-
-class And(val a: Predicate, val b: Predicate) extends Predicate {
-  def isMatch(m: ExecutionContext)(implicit state: QueryState): Option[Boolean] = Ands(List(a, b)).isMatch(m)
-
-  override def atoms: Seq[Predicate] = a.atoms ++ b.atoms
-  override def toString: String = "(" + a + " AND " + b + ")"
-  def containsIsNull = a.containsIsNull||b.containsIsNull
-  def rewrite(f: (Expression) => Expression) = f(And(a.rewriteAsPredicate(f), b.rewriteAsPredicate(f)))
-
-  def arguments = Seq(a, b)
-
-  override def hashCode() = a.hashCode + 37 * b.hashCode
-
-  override def equals(p1: Any) = p1 match {
-    case null       => false
-    case other: And => a == other.a && b == other.b
-    case _          => false
-  }
-
-  def symbolTableDependencies = a.symbolTableDependencies ++ b.symbolTableDependencies
-}
-
-case class Ors(predicates: List[Predicate]) extends CompositeBooleanPredicate {
-  def shouldExitWhen = true
-  def rewrite(f: (Expression) => Expression): Expression = f(Ors(predicates.map(_.rewriteAsPredicate(f))))
-}
-
-case class Or(a: Predicate, b: Predicate) extends Predicate {
-  def isMatch(m: ExecutionContext)(implicit state: QueryState): Option[Boolean] = Ors(List(a, b)).isMatch(m)
-
-  override def toString: String = "(" + a + " OR " + b + ")"
-  def containsIsNull = a.containsIsNull||b.containsIsNull
-  def rewrite(f: (Expression) => Expression) = f(Or(a.rewriteAsPredicate(f), b.rewriteAsPredicate(f)))
-
-  def arguments = Seq(a, b)
-
-  def symbolTableDependencies = a.symbolTableDependencies ++ b.symbolTableDependencies
-}
-
 abstract class CompositeBooleanPredicate extends Predicate {
 
   def predicates: List[Predicate]
